Move resources to a resources sub-package in data, biz and service  - This aligns with the structure used in api/kessel/inventory    to group resources and relationships in different packages
