+++ cmd/serve/serve.go
@@ -4,8 +4,20 @@ import (
 	"context"
 	"fmt"
 	relationshipsctl "github.com/project-kessel/inventory-api/internal/biz/relationships/k8spolicy"
+	hostsctl "github.com/project-kessel/inventory-api/internal/biz/resources/hosts"
+	k8sclustersctl "github.com/project-kessel/inventory-api/internal/biz/resources/k8sclusters"
+	k8spoliciesctl "github.com/project-kessel/inventory-api/internal/biz/resources/k8spolicies"
+	notifsctl "github.com/project-kessel/inventory-api/internal/biz/resources/notificationsintegrations"
 	relationshipsrepo "github.com/project-kessel/inventory-api/internal/data/relationships/k8spolicy"
+	hostsrepo "github.com/project-kessel/inventory-api/internal/data/resources/hosts"
+	k8sclustersrepo "github.com/project-kessel/inventory-api/internal/data/resources/k8sclusters"
+	k8spoliciesrepo "github.com/project-kessel/inventory-api/internal/data/resources/k8spolicies"
+	notifsrepo "github.com/project-kessel/inventory-api/internal/data/resources/notificationsintegrations"
 	relationshipssvc "github.com/project-kessel/inventory-api/internal/service/relationships/k8spolicy"
+	hostssvc "github.com/project-kessel/inventory-api/internal/service/resources/hosts"
+	k8sclusterssvc "github.com/project-kessel/inventory-api/internal/service/resources/k8sclusters"
+	k8spoliciessvc "github.com/project-kessel/inventory-api/internal/service/resources/k8spolicies"
+	notifssvc "github.com/project-kessel/inventory-api/internal/service/resources/notificationsintegrations"
 	"os"
 	"os/signal"
 	"syscall"
@@ -29,20 +41,8 @@ import (
 	pb "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 
 	healthctl "github.com/project-kessel/inventory-api/internal/biz/health"
-	hostsctl "github.com/project-kessel/inventory-api/internal/biz/hosts"
-	k8sclustersctl "github.com/project-kessel/inventory-api/internal/biz/k8sclusters"
-	k8spoliciesctl "github.com/project-kessel/inventory-api/internal/biz/k8spolicies"
-	notifsctl "github.com/project-kessel/inventory-api/internal/biz/notificationsintegrations"
 	healthrepo "github.com/project-kessel/inventory-api/internal/data/health"
-	hostsrepo "github.com/project-kessel/inventory-api/internal/data/hosts"
-	k8sclustersrepo "github.com/project-kessel/inventory-api/internal/data/k8sclusters"
-	k8spoliciesrepo "github.com/project-kessel/inventory-api/internal/data/k8spolicies"
-	notifsrepo "github.com/project-kessel/inventory-api/internal/data/notificationsintegrations"
 	healthssvc "github.com/project-kessel/inventory-api/internal/service/health"
-	hostssvc "github.com/project-kessel/inventory-api/internal/service/hosts"
-	k8sclusterssvc "github.com/project-kessel/inventory-api/internal/service/k8sclusters"
-	k8spoliciessvc "github.com/project-kessel/inventory-api/internal/service/k8spolicies"
-	notifssvc "github.com/project-kessel/inventory-api/internal/service/notificationsintegrations"
 )
 
 func NewCommand(
