+++ internal/service/resources/hosts/hosts.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	pb "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
-	biz "github.com/project-kessel/inventory-api/internal/biz/hosts"
+	biz "github.com/project-kessel/inventory-api/internal/biz/resources/hosts"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	conv "github.com/project-kessel/inventory-api/internal/service/common"
 )
