+++ internal/service/resources/k8sclusters/k8scluster.go
@@ -5,7 +5,7 @@ import (
 	"github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 	pb "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
-	biz "github.com/project-kessel/inventory-api/internal/biz/k8sclusters"
+	biz "github.com/project-kessel/inventory-api/internal/biz/resources/k8sclusters"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	conv "github.com/project-kessel/inventory-api/internal/service/common"
 )
