+++ internal/service/resources/k8spolicies/k8spolicies.go
@@ -6,7 +6,7 @@ import (
 	"github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 	pb "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
-	biz "github.com/project-kessel/inventory-api/internal/biz/k8spolicies"
+	biz "github.com/project-kessel/inventory-api/internal/biz/resources/k8spolicies"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	conv "github.com/project-kessel/inventory-api/internal/service/common"
 )
