+++ internal/service/resources/notificationsintegrations/notificationsintegrations.go
@@ -5,7 +5,7 @@ import (
 
 	pb "github.com/project-kessel/inventory-api/api/kessel/inventory/v1beta1/resources"
 	authnapi "github.com/project-kessel/inventory-api/internal/authn/api"
-	biz "github.com/project-kessel/inventory-api/internal/biz/notificationsintegrations"
+	biz "github.com/project-kessel/inventory-api/internal/biz/resources/notificationsintegrations"
 	"github.com/project-kessel/inventory-api/internal/middleware"
 	conv "github.com/project-kessel/inventory-api/internal/service/common"
 )
